moved logic to new InstanceCalculator class
